PMM-11272 Add API test, add commnet to advisors reponse
