+++ managed/services/checks/checks_test.go
@@ -513,54 +513,6 @@ func TestFilterChecks(t *testing.T) {
 	assert.ElementsMatch(t, expected, actual)
 }
 
-func TestGroupChecksByDB(t *testing.T) {
-	t.Parallel()
-
-	checks := map[string]check.Check{
-		"MySQLShow":                {Name: "MySQLShow", Version: 1, Type: check.MySQLShow},
-		"MySQLSelect":              {Name: "MySQLSelect", Version: 1, Type: check.MySQLSelect},
-		"PostgreSQLShow":           {Name: "PostgreSQLShow", Version: 1, Type: check.PostgreSQLShow},
-		"PostgreSQLSelect":         {Name: "PostgreSQLSelect", Version: 1, Type: check.PostgreSQLSelect},
-		"MongoDBGetParameter":      {Name: "MongoDBGetParameter", Version: 1, Type: check.MongoDBGetParameter},
-		"MongoDBBuildInfo":         {Name: "MongoDBBuildInfo", Version: 1, Type: check.MongoDBBuildInfo},
-		"MongoDBGetCmdLineOpts":    {Name: "MongoDBGetCmdLineOpts", Version: 1, Type: check.MongoDBGetCmdLineOpts},
-		"MongoDBReplSetGetStatus":  {Name: "MongoDBReplSetGetStatus", Version: 1, Type: check.MongoDBReplSetGetStatus},
-		"MongoDBGetDiagnosticData": {Name: "MongoDBGetDiagnosticData", Version: 1, Type: check.MongoDBGetDiagnosticData},
-		"unsupported type":         {Name: "unsupported type", Version: 1, Type: check.Type("RedisInfo")},
-		"missing type":             {Name: "missing type", Version: 1},
-		"MySQL family V2":          {Name: "MySQL family V2", Version: 2, Family: check.MySQL},
-		"PostrgeSQL family V2":     {Name: "PostrgeSQL family V2", Version: 2, Family: check.PostgreSQL},
-		"MongoDB family V2":        {Name: "MongoDB family V2", Version: 2, Family: check.MongoDB},
-		"missing family":           {Name: "missing family", Version: 2},
-	}
-
-	s, err := New(nil, nil, nil, nil, vmAddress)
-	require.NoError(t, err)
-	mySQLChecks, postgreSQLChecks, mongoDBChecks := s.groupChecksByDB(checks)
-
-	require.Len(t, mySQLChecks, 3)
-	require.Len(t, postgreSQLChecks, 3)
-	require.Len(t, mongoDBChecks, 6)
-
-	// V1 checks
-	assert.Equal(t, check.MySQLShow, mySQLChecks["MySQLShow"].Type)
-	assert.Equal(t, check.MySQLSelect, mySQLChecks["MySQLSelect"].Type)
-
-	assert.Equal(t, check.PostgreSQLShow, postgreSQLChecks["PostgreSQLShow"].Type)
-	assert.Equal(t, check.PostgreSQLSelect, postgreSQLChecks["PostgreSQLSelect"].Type)
-
-	assert.Equal(t, check.MongoDBGetParameter, mongoDBChecks["MongoDBGetParameter"].Type)
-	assert.Equal(t, check.MongoDBBuildInfo, mongoDBChecks["MongoDBBuildInfo"].Type)
-	assert.Equal(t, check.MongoDBGetCmdLineOpts, mongoDBChecks["MongoDBGetCmdLineOpts"].Type)
-	assert.Equal(t, check.MongoDBReplSetGetStatus, mongoDBChecks["MongoDBReplSetGetStatus"].Type)
-	assert.Equal(t, check.MongoDBGetDiagnosticData, mongoDBChecks["MongoDBGetDiagnosticData"].Type)
-
-	// V2 checks
-	assert.Equal(t, check.MySQL, mySQLChecks["MySQL family V2"].Family)
-	assert.Equal(t, check.PostgreSQL, postgreSQLChecks["PostrgeSQL family V2"].Family)
-	assert.Equal(t, check.MongoDB, mongoDBChecks["MongoDB family V2"].Family)
-}
-
 func TestMinPMMAgents(t *testing.T) {
 	t.Parallel()
